documentation updated for node_select plugins also change state from an int to an uint16_t

+++ src/plugins/select/bluegene/block_allocator/block_allocator.c
@@ -873,7 +873,7 @@ extern void ba_fini()
  * 
  * IN ba_node: ba_node_t to put down
  */
-extern void ba_update_node_state(ba_node_t *ba_node, int state)
+extern void ba_update_node_state(ba_node_t *ba_node, uint16_t state)
 {
 	uint16_t node_base_state = state & NODE_STATE_BASE;
