+++ src/plugins/select/cons_res/select_cons_res.c
@@ -1342,7 +1342,7 @@ extern int select_p_get_info_from_plugin(enum select_data_info info,
 	return rc;
 }
 
-extern int select_p_update_node_state (int index, int state)
+extern int select_p_update_node_state (int index, uint16_t state)
 {
 	return SLURM_SUCCESS;
 }
